chore: align pre-commit with CI + clear latent clippy warnings
The original cargo fmt fix (c9f695c) only got CI past the rustfmt gate
— the next step (clippy) would have lit up another 39 errors that had
accumulated on dev with no local hook to catch them. Same root cause:
.pre-commit-config.yaml diverged from .github/workflows/ci-*.yml.

Changes:

1. .pre-commit-config.yaml — rewritten to mirror every CI step exactly,
   per the rule "pre-commit hooks must mirror CI checks". Fast checks
   (fmt, clippy --all-targets across all 3 workspaces, including the
   facade --features full variant) run on every commit; slow checks
   (cargo test --workspace, wasm-pack build, maturin build + pytest,
   capi cmake + ctest) run on push. Comments call out the exact CI
   workflow each hook mirrors so future drift is obvious.

2. molrs-io/src/cube.rs — drop 6 redundant `as f64` casts on
   already-f64 origin / cell components in the cube writer.

3. molrs-python — fix 33 clippy errors:
   - 3 redundant `as f64` casts (forcefield.rs, compute.rs)
   - 1 unused `SharedStore` import (store.rs)
   - 5 deprecated `PyAnyMethods::downcast` → `Bound::cast` (block.rs)
   - 12 deprecated automatic FromPyObject for #[pyclass]+Clone types →
     explicit `from_py_object` opt-in (frame.rs PyGrid + molrec.rs all
     6 + compute.rs 5 with #[derive(Clone)])
   - 2 useless `MolRsError → MolRsError` conversions (molrec.rs)
   - 10 too_many_arguments lints in molrec.rs PyO3 bindings — these
     mirror the underlying core API's parameter shape and refactoring
     would change the Python surface, so module-level
     #![allow(clippy::too_many_arguments)] is the correct
     line-of-defense.

4. molrs-wasm/src/io/streaming.rs — one cargo fmt drift the streaming
   WIP commit (5e874d7) introduced; this is what the new
   pre-commit-config would have caught at commit time.

After this commit, `pre-commit run --all-files --hook-stage pre-commit`
is green across all 3 workspaces. The hook framework is now actually
installed (per-repo core.hooksPath override; molrs's .git/hooks were
shadowed by a global hooksPath setting before).

+++ molrs-io/src/cube.rs
@@ -444,9 +444,9 @@ pub fn write_cube_to_writer<W: Write>(writer: &mut W, frame: &Frame) -> Result<(
     // Recover origin and cell columns from simbox, in the writer's unit.
     let origin_arr = simbox.origin_view().to_owned();
     let origin = [
-        origin_arr[0] as f64 * unit_scale,
-        origin_arr[1] as f64 * unit_scale,
-        origin_arr[2] as f64 * unit_scale,
+        origin_arr[0] * unit_scale,
+        origin_arr[1] * unit_scale,
+        origin_arr[2] * unit_scale,
     ];
     let h = simbox.h_view();
 
@@ -478,9 +478,9 @@ pub fn write_cube_to_writer<W: Write>(writer: &mut W, frame: &Frame) -> Result<(
         } else {
             n as i32
         };
-        let vx = (h[[0, i]] as f64 * unit_scale) / n as f64;
-        let vy = (h[[1, i]] as f64 * unit_scale) / n as f64;
-        let vz = (h[[2, i]] as f64 * unit_scale) / n as f64;
+        let vx = (h[[0, i]] * unit_scale) / n as f64;
+        let vy = (h[[1, i]] * unit_scale) / n as f64;
+        let vz = (h[[2, i]] * unit_scale) / n as f64;
         writeln!(writer, "{:5}{:12.6}{:12.6}{:12.6}", n_signed, vx, vy, vz)
             .map_err(MolRsError::Io)?;
     }
